% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_kernel_histogram.R
\name{apply_kernel_histogram}
\alias{apply_kernel_histogram}
\title{apply_kernel_histogram}
\usage{
apply_kernel_histogram(
  fhistogram,
  filter = "circle",
  kernel,
  radius = 2,
  sigma = radius/2
)
}
\arguments{
\item{fhistogram}{Matrix or array interpreted as histogram.}

\item{filter}{Either \code{circle}, \code{square} or \code{gauss}, defaults to \code{circle}.}

\item{kernel}{}

\item{radius}{Used for determining size of kernel, defaults to \code{2}.}

\item{sigma}{Parameter for gaussian filtering, defaults to \code{radius / 2}.}
}
\value{
Blurred histogram with the result.
}
\description{
Blur given histogram using \code{square} or \code{gauss} filtering.
}
