% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blend_rgba_float.R
\name{blend_rgba_float}
\alias{blend_rgba_float}
\title{blend_rgba_float}
\usage{
blend_rgba_float(fRGBA_list)
}
\arguments{
\item{fRGBA_list}{List of floating-point RGBA arrays with premultiplied alpha (each of the same size N-by-M-by-4). The "first" matrix in the list is the one that will be rendered on "top".}
}
\value{
Blended RGBA matrix.
}
\description{
Blend RGBA matrices.
}
