% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorize_data.R
\name{colorize_data}
\alias{colorize_data}
\title{colorize_data}
\usage{
colorize_data(
  xy,
  xlim = c(min(xy[, 1]), max(xy[, 1])),
  ylim = c(min(xy[, 2]), max(xy[, 2])),
  size = c(512, 512),
  rgba = c(0, 0, 0, 255),
  output_raster = TRUE
)
}
\arguments{
\item{xy}{2-column float matrix with point coordinates. As usual with
rasters in R, X axis grows right, and Y axis grows DOWN.
Flipping \code{ylim} causes the "usual" mathematical behavior.}

\item{xlim, }{ylim Float limits as usual (position of the first pixel on the
left/top, and the last pixel on the right/bottom), 2-element vector.
You can easily flip the top/bottom to the "usual" mathematical
system by flipping the \code{ylim} vector.}

\item{size}{2-element vector integer size of the result histogram,
defaults to \code{c(512,512)}.}

\item{rgba}{vector with 4 elements or matrix or array (4xn dim, n >= 2, n ~ xy rows) with R, G, B
and alpha channels  in integers, defaults to \code{c(0,0,0,255)}}

\item{output_raster}{If the returned result is in raster form, defaults to \code{TRUE}. \code{FALSE}
for performing other operations.}
}
\value{
Raster or integer matrix with the result.
}
\description{
Colorize given data with one given color or each point has its corresponding given color.
}
