% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorize_hist.R
\name{colorize_hist}
\alias{colorize_hist}
\title{colorize_hist}
\usage{
colorize_hist(
  hist,
  rgba = array(c(250, 128, 114, 255, 144, 238, 144, 255, 176, 224, 230, 255), c(4, 3)),
  output_raster = TRUE
)
}
\arguments{
\item{hist}{Matrix or array R datatype interpreted as histogram.}

\item{rgba}{Matrix (4xn dim, n>= 2) with R, G, B and alpha channels
in integers, defaults to shades of \code{red}, \code{green} and \code{blue} with \code{alpha = 255}.}

\item{output_raster}{If the returned result is in raster form, defaults to \code{TRUE}. \code{FALSE}
for performing other operations.}
}
\value{
Raster or integer matrix with the result.
}
\description{
Colorize given histogram with input palette.
}
