% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorize_hist.R
\name{colorize_histogram}
\alias{colorize_histogram}
\title{colorize_histogram}
\usage{
colorize_histogram(
  fhistogram,
  RGBA = array(c(250, 128, 114, 255, 144, 238, 144, 255, 176, 224, 230, 255), c(4, 3))
)
}
\arguments{
\item{RGBA}{Integer matrix (4xn dim, n>= 2) with R, G, B and alpha channels
in integers, defaults to shades of \code{red}, \code{green} and \code{blue} with \code{alpha = 255}.}

\item{histogram}{Matrix or array R datatype interpreted as histogram.}
}
\value{
Float RGBWT matrix.
}
\description{
Colorize given histogram with input palette.
}
