% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scattermore.R
\docType{data}
\name{GeomScattermore}
\alias{GeomScattermore}
\title{The actual geom for scattermore}
\format{
An object of class \code{GeomScattermore} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomScattermore
}
\description{
The actual geom for scattermore
}
\keyword{datasets}
