% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scattermore.R
\docType{data}
\name{GeomScattermost}
\alias{GeomScattermost}
\title{The actual geom for scattermost}
\format{
An object of class \code{GeomScattermost} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
GeomScattermost
}
\description{
The actual geom for scattermost
}
\keyword{datasets}
