% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_to_rgbwt.R
\name{histogram_to_rgbwt}
\alias{histogram_to_rgbwt}
\title{histogram_to_rgbwt}
\usage{
histogram_to_rgbwt(
  fhistogram,
  RGBA = grDevices::col2rgb(col, alpha = T),
  col = grDevices::hcl.colors(10),
  zlim = c(min(fhistogram), max(fhistogram))
)
}
\arguments{
\item{fhistogram}{Matrix or 2D array with the histogram of values.}

\item{RGBA}{4-by-N matrix floating-point R, G, B and A channels for the palette. Overrides \code{col}.}

\item{col}{Colors to use for coloring.}

\item{zlim}{Values to use as extreme values of the histogram}
}
\value{
RGBWT matrix.
}
\description{
Colorize given histogram with input palette.
}
