% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_raster.R
\name{make_raster}
\alias{make_raster}
\title{make_raster}
\usage{
make_raster(
  xy,
  xlim = c(min(xy[, 1]), max(xy[, 1])),
  ylim = c(min(xy[, 2]), max(xy[, 2])),
  size = c(512, 512),
  rgba = c(0, 0, 0, 255)
)
}
\arguments{
\item{xy}{2-column float matrix with point coordinates. As usual with
rasters in R, X axis grows right, and Y axis grows DOWN.
Flipping \code{ylim} causes the "usual" mathematical behavior.}

\item{xlim, }{ylim Float limits as usual (position of the first pixel on the
left/top, and the last pixel on the right/bottom), 2-element vector.
You can easily flip the top/bottom to the "usual" mathematical
system by flipping the \code{ylim} vector.}

\item{size}{2-element vector integer size of the result histogram,
defaults to \code{c(512,512)}.}

\item{rgba}{vector with 4 elements or matrix (4xn dim, n >= 2, n ~ xy rows) with R, G, B
and alpha channels  in integers, defaults to \code{c(0,0,0,255)}}
}
\value{
float matrix with the result.
}
\description{
Create raster from given data.
}
