% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rgbwt.R
\name{merge_rgbwt}
\alias{merge_rgbwt}
\title{merge_rgbwt}
\usage{
merge_rgbwt(fRGBWT_list)
}
\arguments{
\item{fRGBWT_list}{List of RGBWT arrays. The order of the matrices does not matter (except for negligible floating-point rounding and other robustness errors).}
}
\value{
Merged RGBWT matrix.
}
\description{
Merge RGBWT matrices.
}
