% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgba_float_to_rgba_int.R
\name{rgba_float_to_rgba_int}
\alias{rgba_float_to_rgba_int}
\title{rgba_float_to_rgba_int}
\usage{
rgba_float_to_rgba_int(fRGBA)
}
\arguments{
\item{fRGBA}{RGBA matrix (\code{red}, \code{green}, \code{blue} and \code{alpha} channels, dimension nxmx4, values ~ 0-1).}
}
\value{
RGBA matrix, output \emph{is not premultiplied} by alpha.
}
\description{
Convert RGBA matrix to RGBA matrix.
}
