% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgba_int_to_raster.R
\name{rgba_int_to_raster}
\alias{rgba_int_to_raster}
\title{rgba_int_to_raster}
\usage{
rgba_int_to_raster(i32RGBA)
}
\arguments{
\item{i32RGBA}{Integer RGBA matrix (\code{red}, \code{green}, \code{blue} and \code{alpha} channels, dimension nxmx4, values ~ 0-255).}
}
\value{
Raster result.
}
\description{
Create raster from given RGBA matrix.
}
