% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgba_to_raster.R
\name{rgba_to_raster}
\alias{rgba_to_raster}
\title{rgba_to_raster}
\usage{
rgba_to_raster(rgba)
}
\arguments{
\item{rgba}{Integer rgba matrix (\code{red}, \code{green}, \code{blue} and \code{alpha} channels, dimension nxmx4).}
}
\value{
Raster result.
}
\description{
Create raster from given rgba matrix.
}
