% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgbwt_to_raster.R
\name{rgbwt_to_raster}
\alias{rgbwt_to_raster}
\title{rgbwt_to_raster}
\usage{
rgbwt_to_raster(rgbwt)
}
\arguments{
\item{rgbwt}{Integer rgbwt matrix (\code{red}, \code{green}, \code{blue} channels, \code{weight} ~ sum of alphas,
\code{transparency} ~ 1 - alpha, dimension nxmx5).}
}
\value{
Raster result.
}
\description{
Create raster from given rgbwt matrix.
}
