% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgbwt_to_rgba.R
\name{rgbwt_to_rgba}
\alias{rgbwt_to_rgba}
\title{rgbwt_to_rgba}
\usage{
rgbwt_to_rgba(rgbwt)
}
\arguments{
\item{rgbwt}{Integer rgbwt matrix (\code{red}, \code{green}, \code{blue} channels, \code{weight} ~ sum of alphas,
\code{transparency} ~ 1 - alpha, dimension nxmx5).}
}
\value{
Integer rgba matrix.
}
\description{
Convert rgbwt matrix to rgba matrix.
}
