% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgbwt_to_rgba_float.R
\name{rgbwt_to_rgba_float}
\alias{rgbwt_to_rgba_float}
\title{rgbwt_to_rgba_float}
\usage{
rgbwt_to_rgba_float(fRGBWT)
}
\arguments{
\item{fRGBWT}{The RGBWT matrix.}
}
\value{
RGBA matrix, output \emph{is premultiplied} by alpha.
}
\description{
Convert RGBWT matrix to floating-point RGBA matrix, suitable for alpha-blending.
}
