% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scattermore.R
\name{scattermore}
\alias{scattermore}
\title{scattermore}
\usage{
scattermore(xy, size = c(512, 512), xlim = c(min(xy[, 1]), max(xy[,
  1])), ylim = c(min(xy[, 2]), max(xy[, 2])), rgba = c(0L, 0L, 0L,
  255L), cex = 0)
}
\arguments{
\item{xy}{2-column float matrix with point coordinates. As usual with
rasters in R, 'x' axis grows right, and 'y' axis grows DOWN.
Flipping `ylim` causes the usual mathematical behavior.}

\item{size}{2-element vector integer size of the result raster,
defaults to 512x512.}

\item{xlim, ylim}{Float limits as usual (position of the first pixel on the
left/top, and the last pixel on the right/bottom). You can
easily flip the top/bottom to the "usual" mathematical
system by flipping the ylim vector.}

\item{rgba}{4-row matrix with color values of 0-255, or just a single 4-item
vector for c(r,g,b,a). Best created with col2rgb(..., alpha=T).}

\item{cex}{Point radius in pixels, 0=single-pixel dots.}
}
\value{
Raster with the result.
}
\description{
Convert points to raster scatterplot rather quickly.
}
\details{
scattermore::scattermore(cbind(rnorm(1000),rnorm(1000)))
}
