% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_kernel_data.R
\name{apply_kernel_data}
\alias{apply_kernel_data}
\title{apply_kernel_data}
\usage{
apply_kernel_data(
  rgbwt,
  rgba = NULL,
  filter = "circle",
  radius = 5,
  sigma = 10,
  output_raster = TRUE
)
}
\arguments{
\item{rgbwt}{Integer rgbwt matrix (\code{red}, \code{green}, \code{blue} channels, \code{weight} ~ sum of alphas,
\code{transparency} ~ 1 - alpha).}

\item{filter}{Either \code{circle} or \code{gaussian} (symmetric).}

\item{radius}{Size of circle kernel (float), defaults to \code{5}.}

\item{sigma}{Parameter for gaussian filtering, defaults to \code{10}.}
}
\value{
Raster result or list with integer rgbwt and rgba matrices.
}
\description{
Blur given rgbwt matrix using \code{circle} or \code{gauss} filtering.
}
