% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_kernel_data.R
\name{apply_kernel_data}
\alias{apply_kernel_data}
\title{apply_kernel_data}
\usage{
apply_kernel_data(
  rgbwt,
  filter = "circle",
  kernel_pixels = 2,
  sigma = 10,
  output_raster = TRUE
)
}
\arguments{
\item{filter}{Either \code{circle} or \code{gaussian} (symmetric).}

\item{kernel_pixels}{Used for determining size of kernel,
(\code{size = 2*kernel_pixels + 1}), defaults to \code{2}.}

\item{sigma}{Parameter for gaussian filtering, defaults to \code{10}.}

\item{hist}{Integer rgbwt matrix (\code{red}, \code{green}, \code{blue} channels, \code{weight} ~ sum of alphas,
\code{transparency} ~ 1 - alpha).}
}
\value{
Raster or integer matrix with the resul
}
\description{
Blur given rgbwt matrix using \code{circle} or \code{gauss} filtering.
}
