% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_kernel_data.R
\name{apply_kernel_data}
\alias{apply_kernel_data}
\title{apply_kernel_data}
\usage{
apply_kernel_data(
  fRGBWT,
  filter = "circle",
  radius = 2,
  sigma = 1,
  approx_limit = 3.5
)
}
\arguments{
\item{fRGBWT}{Float RGBWT matrix (\code{red}, \code{green}, \code{blue} channels, \code{weight} ~ sum of alphas,
\code{transparency} ~ 1 - alpha, dimension nxmx5).}

\item{filter}{Either \code{circle} or \code{gaussian} (symmetric).}

\item{radius}{Size of circle kernel (float), defaults to \code{2}.}

\item{sigma}{Parameter for gaussian filtering, defaults to \code{1}.}

\item{approx_limit}{Sets the size of the kernel square, multiplied with \code{sigma}, defaults to \code{3.5}.}
}
\value{
Float RGBWT matrix.
}
\description{
Blur given RGBWT matrix using \code{circle} or \code{gauss} filtering.
}
