% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_kernel_hist.R
\name{apply_kernel_hist}
\alias{apply_kernel_hist}
\title{apply_kernel_hist}
\usage{
apply_kernel_hist(hist, filter = "classic", size = 5, sigma = 1)
}
\arguments{
\item{hist}{float matrix or array R datatype interpreted as histogram}

\item{filter}{either \code{classic}(matrix of ones) or \code{gaussian}}

\item{size}{dimension of kernel that is used, changed to odd it not,
defaults to \code{5}.}

\item{sigma}{parameter for gaussian filtering, defaults to \code{1.0}}
}
\value{
float matrix with the result.
}
\description{
Blur given histogram using \code{classic} or \code{gauss} filtering.
}
