% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_kernel_histogram.R
\name{apply_kernel_histogram}
\alias{apply_kernel_histogram}
\title{apply_kernel_histogram}
\usage{
apply_kernel_histogram(
  fhistogram,
  filter = "circle",
  mask,
  radius = 2,
  sigma = radius/2,
  threads = 1
)
}
\arguments{
\item{fhistogram}{Matrix or array interpreted as histogram.}

\item{filter}{Either \code{circle}, \code{square}, \code{gauss} or \code{own}, defaults to \code{circle}.}

\item{mask}{Own kernel used for blurring, \code{filter} parameter has to be set to \code{own}.}

\item{radius}{Used for determining size of kernel, defaults to \code{2}.}

\item{sigma}{Parameter for gaussian filtering, defaults to \code{radius / 2}.}

\item{threads}{Ensures multithreading when blurring the histogram, defaults to \code{1} (no parallelization).}
}
\value{
Blurred histogram with the result.
}
\description{
Blur given histogram using \code{square} or \code{gauss} filtering.
}
