% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blend_rgba_float.R
\name{blend_rgba_float}
\alias{blend_rgba_float}
\title{blend_rgba_float}
\usage{
blend_rgba_float(fRGBA_list)
}
\arguments{
\item{fRGBA_list}{list of RGBA matrices (\code{red}, \code{green}, \code{blue} and \code{alpha} channels, dimension nxmx4, values ~ 0-1).}
}
\value{
RGBA matrix.
}
\description{
Blend RGBA matrices.
}
