% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorize_hist.R
\name{colorize_hist}
\alias{colorize_hist}
\title{colorize_hist}
\usage{
colorize_hist(
  hist,
  palette = array(c(0, 0, 0, 255, 255, 255, 255, 255), c(4, 2))
)
}
\arguments{
\item{hist}{matrix or array R datatype interpreted as histogram}

\item{pallete}{matrix (4xn dim, n>= 2) with R, G, B and alpha channels
in integers, defaults to \code{black} and \code{white} with \code{alpha = 255}}
}
\value{
float matrix with the result.
}
\description{
Colorize given histogram with input palette.
}
