% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_to_rgbwt.R
\name{histogram_to_rgbwt}
\alias{histogram_to_rgbwt}
\title{histogram_to_rgbwt}
\usage{
histogram_to_rgbwt(
  fhistogram,
  RGBA = array(c(250, 128, 114, 255, 144, 238, 144, 255, 176, 224, 230, 255), c(4, 3))
)
}
\arguments{
\item{fhistogram}{Matrix or array R datatype interpreted as histogram.}

\item{RGBA}{matrix (4xn dim, n>= 2) with R, G, B and alpha channels
in integers, defaults to shades of \code{red}, \code{green} and \code{blue} with \code{alpha = 255}.}
}
\value{
RGBWT matrix.
}
\description{
Colorize given histogram with input palette.
}
