% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rgbwt.R
\name{merge_rgbwt}
\alias{merge_rgbwt}
\title{merge_rgbwt}
\usage{
merge_rgbwt(fRGBWT_1, fRGBWT_2)
}
\arguments{
\item{fRGBWT_1}{Float RGBWT matrix (\code{red}, \code{green}, \code{blue} channels, \code{weight} ~ sum of alphas,
\code{transparency} ~ 1 - alpha, dimension nxmx5).}

\item{fRGBWT_2}{The same as fRGBWT_1 parameter.}
}
\value{
Float RGBWT matrix.
}
\description{
Blend two RGBWT matrices.
}
