% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rgbwt.R
\name{merge_rgbwt}
\alias{merge_rgbwt}
\title{merge_rgbwt}
\usage{
merge_rgbwt(fRGBWT_list)
}
\arguments{
\item{fRGBWT_list}{list of RGBWT matrices (\code{red}, \code{green}, \code{blue} channels, \code{weight} ~ sum of alphas,
\code{transparency} ~ 1 - alpha, dimension nxmx5).}
}
\value{
RGBWT matrix.
}
\description{
Merge RGBWT matrices.
}
