% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgba_float_to_rgba_int.R
\name{rgba_float_to_rgba_int}
\alias{rgba_float_to_rgba_int}
\title{rgba_float_to_rgba_int}
\usage{
rgba_float_to_rgba_int(fRGBA)
}
\arguments{
\item{fRGBA}{RGBA bitmap in N-by-M-by-4 array.}
}
\value{
RGBA matrix. The output \emph{is not premultiplied} by alpha.
}
\description{
Convert a float RGBA bitmap with pre-multiplied alpha to integer RGBA bitmap.
}
