% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgba_int_to_raster.R
\name{rgba_int_to_raster}
\alias{rgba_int_to_raster}
\title{rgba_int_to_raster}
\usage{
rgba_int_to_raster(i32RGBA)
}
\arguments{
\item{i32RGBA}{Integer RGBA matrix (with all values between 0 and 255).}
}
\value{
The matrix converted to raster.
}
\description{
Create a raster from the given RGBA matrix.
}
