% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgbwt_to_rgba_float.R
\name{rgbwt_to_rgba_float}
\alias{rgbwt_to_rgba_float}
\title{rgbwt_to_rgba_float}
\usage{
rgbwt_to_rgba_float(fRGBWT)
}
\arguments{
\item{fRGBWT}{Float rgbwt matrix (\code{red}, \code{green}, \code{blue} channels, \code{weight} ~ sum of alphas,
\code{transparency} ~ 1 - alpha, dimension nxmx5).}
}
\value{
Float RGBA matrix.
}
\description{
Convert rgbwt matrix to float RGBA matrix.
}
