% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgbwt_to_rgba_int.R
\name{rgbwt_to_rgba_int}
\alias{rgbwt_to_rgba_int}
\title{rgbwt_to_rgba_int}
\usage{
rgbwt_to_rgba_int(fRGBWT)
}
\arguments{
\item{fRGBWT}{The RGBWT matrix.}
}
\value{
A RGBA matrix. The output \emph{is not premultiplied} by alpha.
}
\description{
Convert a RGBWT matrix to an integer RGBA matrix.
}
