% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgbwt_to_rgba_int.R
\name{rgbwt_to_rgba_int}
\alias{rgbwt_to_rgba_int}
\title{rgbwt_to_rgba_int}
\usage{
rgbwt_to_rgba_int(rgbwt)
}
\arguments{
\item{rgbwt}{float rgbwt matrix (\code{red}, \code{green}, \code{blue} channels, \code{weight} ~ sum of alphas,
\code{transparency} ~ 1 - alpha, dimension nxmx5).}
}
\value{
Integer rgba matrix.
}
\description{
Convert rgbwt matrix to integer rgba matrix.
}
