% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scattermore.R
\name{scattermoreplot}
\alias{scattermoreplot}
\title{scattermoreplot}
\usage{
scattermoreplot(x, y, xlim, ylim, size, col = rgb(0, 0, 0, 1), cex = 0,
  ...)
}
\arguments{
\item{x, y, xlim, ylim, ...}{used as in plot() or forwarded to plot()}

\item{size}{forwarded to scattermore(), or auto-derived from device and plot size if missing (the estimate is not pixel-perfect, but pretty close)}

\item{col}{point color(s)}

\item{cex}{forwarded to scattermore()}
}
\description{
Convenience base-graphics-like layer around scattermore. Currently only works with linear axes!
}
\examples{
library(scattermore)
scattermoreplot(
  rnorm(1e7),
  rnorm(1e7),
  col=heat.colors(1e7, alpha=.1),
  main="scattermore demo")
}
